+++ ring/lifecycler_metrics.go
@@ -19,16 +19,6 @@ func NewLifecyclerMetrics(ringName string, reg prometheus.Registerer) *Lifecycle
 			Help:        "The total number of heartbeats sent to consul.",
 			ConstLabels: prometheus.Labels{"name": ringName},
 		}),
-		tokensOwned: promauto.With(reg).NewGauge(prometheus.GaugeOpts{
-			Name:        "member_ring_tokens_owned",
-			Help:        "The number of tokens owned in the ring.",
-			ConstLabels: prometheus.Labels{"name": ringName},
-		}),
-		tokensToOwn: promauto.With(reg).NewGauge(prometheus.GaugeOpts{
-			Name:        "member_ring_tokens_to_own",
-			Help:        "The number of tokens to own in the ring.",
-			ConstLabels: prometheus.Labels{"name": ringName},
-		}),
 		shutdownDuration: promauto.With(reg).NewHistogramVec(prometheus.HistogramOpts{
 			Name:        "shutdown_duration_seconds",
 			Help:        "Duration (in seconds) of shutdown procedure (ie transfer or flush).",